Move WorkerTaskObject to separate file
